Factor out yaml processing out of config_manage
- Create new utility module for processing yaml
- Move yaml processing into new utility module
- Factor out Schema classes out of config_manage into schema module
- Refactor yaml reading to accommodate a unit test (minor change)
- Add unit test for new schema module

Overall: this is first step towards refactoring config.
